Change etag to be the local etag rather than the remote.
We were exposing the remote etag, that is, the etag of the item when it
was last synced, though there's really no reason why anyone would want
that.
The etag is supposed to be based on the content of the item, in its
current state, so it can actually be used, and not some internal data.