CI: use sccache for caching rust builds.
This fixes https://github.com/etesync/etesync-dav/issues/165

The issue was that we were caching too aggressively, combined with that
on Windows (unlike Linux and macOS) it was linking the module against
the python library. The aggressive caching meant that all of the builds
were linking against the Python 3.6 DLL, making it not work with other
Python versions.